+++ src/blocks.js
@@ -160,7 +160,7 @@ CustomCommandBlockMorph, ToggleButtonMorph, DialMorph, SnapExtensions*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.blocks = '2021-September-29';
+modules.blocks = '2021-September-30';
 
 var SyntaxElementMorph;
 var BlockMorph;
@@ -983,6 +983,7 @@ SyntaxElementMorph.prototype.labelParts = {
         label: (optional)
         tags: 'widget' // doesn't count as "empty" slot implicit parameter
         min: (optional) number of minimum inputs) or zero
+        max: (optional) number of maximum inputs) or zero
         defaults: (optional) number of visible slots to begin with or zero
     */
     '%inputs': {
@@ -990,6 +991,13 @@ SyntaxElementMorph.prototype.labelParts = {
         slots: '%s',
         label: 'with inputs',
         tags: 'widget'
+    },
+    '%send': {
+        type: 'multi',
+        slots: '%s',
+        label: 'and send',
+        tags: 'static',
+        max: 1
     },
     '%scriptVars': {
         type: 'multi',
@@ -1002,6 +1010,18 @@ SyntaxElementMorph.prototype.labelParts = {
         slots: '%t',
         label: 'block variables',
         tags: 'widget'
+    },
+    '%message': {
+        type: 'multi',
+        slots: '%t',
+        tags: 'widget',
+        max: 1
+    },
+    '%keyName': {
+        type: 'multi',
+        slots: '%t',
+        tags: 'widget',
+        max: 1
     },
     '%parms': {
         type: 'multi',
@@ -1654,6 +1674,7 @@ SyntaxElementMorph.prototype.labelPart = function (spec) {
                 info.min || 0,
                 spec
             );
+            part.maxInputs = info.max;
             for (i = 0; i < info.defaults || 0; i += 1) {
                 part.addInput();
             }
@@ -11599,6 +11620,7 @@ MultiArgMorph.prototype.init = function (
     this.slotSpec = slotSpec || '%s';
     this.labelText = localize(labelTxt || '');
     this.minInputs = min || 0;
+    this.maxInputs = null;
     this.elementSpec = eSpec || null;
     this.labelColor = labelColor || null;
     this.shadowColor = shadowColor || null;
@@ -11748,6 +11770,8 @@ MultiArgMorph.prototype.fixArrowsLayout = function () {
         rightArrow = arrows.children[1],
         inpCount = this.inputs().length,
         dim = new Point(rightArrow.width() / 2, rightArrow.height());
+    leftArrow.show();
+    rightArrow.show();
     if (inpCount < (this.minInputs + 1)) { // hide left arrow
         if (label) {
             label.hide();
@@ -11768,6 +11792,11 @@ MultiArgMorph.prototype.fixArrowsLayout = function () {
         rightArrow.show();
         rightArrow.setPosition(leftArrow.topCenter());
         arrows.bounds.corner = rightArrow.bottomRight().copy();
+        if (!isNil(this.maxInputs) && inpCount > this.maxInputs - 1) {
+            // hide right arrow
+            rightArrow.hide();
+            arrows.setExtent(dim);
+        }
     }
     arrows.rerender();
 };
@@ -11826,6 +11855,10 @@ MultiArgMorph.prototype.addInput = function (contents) {
         } else {
             newPart.setContents('#' + idx);
         }
+    } else if (this.elementSpec === '%message') {
+        newPart.setContents(localize('message'));
+    } else if (this.elementSpec === '%keyName') {
+        newPart.setContents(localize('key'));
     }
     newPart.parent = this;
     this.children.splice(idx, 0, newPart);